link-saver/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
beautifulsoup4
tldextract
passlib[bcrypt]
python-jose[cryptography]
google-generativeai
